+++ bin/utility/office-windows/update_g00.bat
@@ -3,10 +3,10 @@ cd d:\vob
 cd d:\w090\archive\g00
 mkdir %date:~0,10%
 cd %date:~0,10%
-xcopy /s /y d:\vob\g00
+if exist d:\vob\g00  xcopy /s /y d:\vob\g00
 cd d:\vob
-rm -rf g00
+if exist d:\vob\g00 rm -rf g00
 set http_proxy=wwwgate0-ch.mot.com:1080
-git clone http://grasp:improvew090#@github.com/grasp/g00.git
+git clone http://grasp:@github.com/grasp/g00.git
 pause
